chore: move redteam command locations
