+++ src/commands/redteam/generate.ts
