+++ src/onboarding.ts
@@ -3,7 +3,7 @@ import rawlist from '@inquirer/rawlist';
 import chalk from 'chalk';
 import fs from 'fs';
 import path from 'path';
-import { redteamInit } from './commands/redteam';
+import { redteamInit } from './commands/redteam/init';
 import { getEnvString } from './envars';
 import logger from './logger';
 import telemetry, { type EventValue } from './telemetry';
