+++ test/commands/redteam/generate.test.ts
@@ -1,5 +1,5 @@
 import fs from 'fs';
-import { doGenerateRedteam } from '../../../src/commands/generate/redteam';
+import { doGenerateRedteam } from '../../../src/commands/redteam/generate';
 import * as configModule from '../../../src/config';
 import logger from '../../../src/logger';
 import { synthesize } from '../../../src/redteam';
